Fixes a typo (emitVariableDeclarations) Refactorings to address comments on the review (TextureBaker, MetalTexture class, Math-Scalar Operators, MSLProgram, ...)
