--- libraries/stdlib/genmsl/lib/mx_transform_uv.metal
@@ -1,5 +1,0 @@
-vec2 mx_transform_uv(vec2 uv, vec2 uv_scale, vec2 uv_offset)
-{
-    uv = uv * uv_scale + uv_offset;
-    return uv;
-}
