+++ source/MaterialXGenMsl/MslResourceBindingContext.cpp
@@ -71,12 +71,15 @@ void MslResourceBindingContext::emitStructuredResourceBindings(GenContext& conte
     ShaderGenerator& generator = context.getShaderGenerator();
 
     const size_t baseAlignment = 16;
-    std::unordered_map<const TypeDesc*, size_t> alignmentMap({ { Type::FLOAT, baseAlignment / 4 },
+    // Values are adjusted based on
+    // https://developer.apple.com/metal/Metal-Shading-Language-Specification.pdf
+    std::unordered_map<const TypeDesc*, size_t> alignmentMap({
+        { Type::FLOAT, baseAlignment / 4 },
         { Type::INTEGER, baseAlignment / 4 },
         { Type::BOOLEAN, baseAlignment / 4 },
         { Type::COLOR3, baseAlignment },
         { Type::COLOR4, baseAlignment },
-        { Type::VECTOR2, baseAlignment },
+        { Type::VECTOR2, baseAlignment / 2 },
         { Type::VECTOR3, baseAlignment },
         { Type::VECTOR4, baseAlignment },
         { Type::MATRIX33, baseAlignment * 4 },
