+++ source/MaterialXGenMsl/MslShaderGenerator.h
@@ -13,6 +13,9 @@
 
 #include <MaterialXGenShader/HwShaderGenerator.h>
 
+#define TEXTURE_NAME(t) (t + "_tex")
+#define SAMPLER_NAME(t) (t + "_sampler")
+
 MATERIALX_NAMESPACE_BEGIN
 
 using MslShaderGeneratorPtr = shared_ptr<class MslShaderGenerator>;
@@ -58,17 +61,17 @@ class MX_GENMSL_API MslShaderGenerator : public HwShaderGenerator
     virtual void emitVertexStage(const ShaderGraph& graph, GenContext& context, ShaderStage& stage) const;
     virtual void emitPixelStage(const ShaderGraph& graph, GenContext& context, ShaderStage& stage) const;
 
-    virtual void emitMetalTextureClass(ShaderStage& stage) const;
+    virtual void emitMetalTextureClass(GenContext& context, ShaderStage& stage) const;
     virtual void emitDirectives(GenContext& context, ShaderStage& stage) const;
     virtual void emitConstants(GenContext& context, ShaderStage& stage) const;
     virtual void emitLightData(GenContext& context, ShaderStage& stage) const;
     virtual void emitInputs(GenContext& context, ShaderStage& stage) const;
     virtual void emitOutputs(GenContext& context, ShaderStage& stage) const;
     
-    virtual string MathMatrixScalarMathOperators() const;
+    virtual void emitMathMatrixScalarMathOperators(GenContext& context, ShaderStage& stage) const;
     virtual void MetalizeGeneratedShader(ShaderStage& shaderStage) const;
  
-    void emitConstantBufferDeclerations(GenContext& context,
+    void emitConstantBufferDeclarations(GenContext& context,
                                         HwResourceBindingContextPtr resourceBindingCtx,
                                         ShaderStage& stage) const;
