+++ source/MaterialXRender/Material.cpp
@@ -4,6 +4,7 @@
 //
 
 #include <MaterialXRender/Material.h>
+#include <MaterialXFormat/XmlIo.h>
 
 MATERIALX_NAMESPACE_BEGIN
 
@@ -58,6 +59,45 @@ ShaderPtr Material::getShader() const
 bool Material::hasTransparency() const
 {
     return _hasTransparency;
+}
+
+bool Material::generateEnvironmentShader(GenContext& context,
+                                         const FilePath& filename,
+                                         DocumentPtr stdLib,
+                                         const FilePath& imagePath)
+{
+    // Read in the environment nodegraph.
+    DocumentPtr doc = createDocument();
+    doc->importLibrary(stdLib);
+    DocumentPtr envDoc = createDocument();
+    readFromXmlFile(envDoc, filename);
+    doc->importLibrary(envDoc);
+
+    NodeGraphPtr envGraph = doc->getNodeGraph("environmentDraw");
+    if (!envGraph)
+    {
+        return false;
+    }
+    NodePtr image = envGraph->getNode("envImage");
+    if (!image)
+    {
+        return false;
+    }
+    image->setInputValue("file", imagePath.asString(), FILENAME_TYPE_STRING);
+    OutputPtr output = envGraph->getOutput("out");
+    if (!output)
+    {
+        return false;
+    }
+
+    // Create the shader.
+    std::string shaderName = "__ENV_SHADER__";
+    _hwShader = createShader(shaderName, context, output);
+    if (!_hwShader)
+    {
+        return false;
+    }
+    return generateShader(_hwShader);
 }
 
 MATERIALX_NAMESPACE_END
