+++ source/MaterialXRenderMsl/MetalState.mm
@@ -128,7 +128,9 @@ void MetalState::initLinearToSRGBKernel()
          "{                                                             \n"
          "    half4 color_srgb;                                         \n"
          "    for(int i = 0; i < 3; ++i)                                \n"
-         "        color_srgb[i] = (color_linear[i] < 0.0031308) ? (12.92 * color_linear[i]): (1.055 * pow(color_linear[i], 1.0h / 2.2h) - 0.055);\n"
+         "        color_srgb[i] = (color_linear[i] < 0.0031308) ?       \n"
+         "          (12.92 * color_linear[i])                   :       \n"
+         "          (1.055 * pow(color_linear[i], 1.0h / 2.2h) - 0.055);\n"
          "    color_srgb[3] = color_linear[3];                          \n"
          "    return color_srgb;                                        \n"
          "}                                                             \n"
