+++ source/MaterialXRenderMsl/MslMaterial.mm
@@ -94,45 +94,6 @@ bool MslMaterial::generateShader(ShaderPtr hwShader)
     return true;
 }
 
-bool MslMaterial::generateEnvironmentShader(GenContext& context,
-                                         const FilePath& filename,
-                                         DocumentPtr stdLib,
-                                         const FilePath& imagePath)
-{
-    // Read in the environment nodegraph. 
-    DocumentPtr doc = createDocument();
-    doc->importLibrary(stdLib);
-    DocumentPtr envDoc = createDocument();
-    readFromXmlFile(envDoc, filename);
-    doc->importLibrary(envDoc);
-
-    NodeGraphPtr envGraph = doc->getNodeGraph("environmentDraw");
-    if (!envGraph)
-    {
-        return false;
-    }
-    NodePtr image = envGraph->getNode("envImage");
-    if (!image)
-    {
-        return false;
-    }
-    image->setInputValue("file", imagePath.asString(), FILENAME_TYPE_STRING);
-    OutputPtr output = envGraph->getOutput("out");
-    if (!output)
-    {
-        return false;
-    }
-
-    // Create the shader.
-    std::string shaderName = "__ENV_SHADER__";
-    _hwShader = createShader(shaderName, context, output); 
-    if (!_hwShader)
-    {
-        return false;
-    }
-    return generateShader(_hwShader);
-}
-
 bool MslMaterial::bindShader() const
 {
     if (_glProgram)
@@ -283,7 +244,7 @@ void MslMaterial::bindLighting(LightHandlerPtr lightHandler,
     _glProgram->bindLighting(lightHandler, imageHandler);
 
     // Bind shadow map properties
-    if (shadowState.shadowMap && _glProgram->hasUniform(HW::SHADOW_MAP + "_tex"))
+    if (shadowState.shadowMap && _glProgram->hasUniform(TEXTURE_NAME(HW::SHADOW_MAP)))
     {
         ImageSamplingProperties samplingProperties;
         samplingProperties.uaddressMode = ImageSamplingProperties::AddressMode::CLAMP;
@@ -291,12 +252,12 @@ void MslMaterial::bindLighting(LightHandlerPtr lightHandler,
         samplingProperties.filterType = ImageSamplingProperties::FilterType::LINEAR;
 
         // Bind the shadow map.
-        _glProgram->bindTexture(imageHandler, HW::SHADOW_MAP + "_tex", shadowState.shadowMap, samplingProperties);
+        _glProgram->bindTexture(imageHandler, TEXTURE_NAME(HW::SHADOW_MAP), shadowState.shadowMap, samplingProperties);
         _glProgram->bindUniform(HW::SHADOW_MATRIX, Value::createValue(shadowState.shadowMatrix));
     }
 
     // Bind ambient occlusion properties.
-    if (shadowState.ambientOcclusionMap && _glProgram->hasUniform(HW::AMB_OCC_MAP + "_tex"))
+    if (shadowState.ambientOcclusionMap && _glProgram->hasUniform(TEXTURE_NAME(HW::AMB_OCC_MAP)))
     {
         ImageSamplingProperties samplingProperties;
         samplingProperties.uaddressMode = ImageSamplingProperties::AddressMode::PERIODIC;
@@ -304,7 +265,7 @@ void MslMaterial::bindLighting(LightHandlerPtr lightHandler,
         samplingProperties.filterType = ImageSamplingProperties::FilterType::LINEAR;
 
         // Bind the ambient occlusion map.
-        _glProgram->bindTexture(imageHandler, HW::AMB_OCC_MAP + "_tex",
+        _glProgram->bindTexture(imageHandler, TEXTURE_NAME(HW::AMB_OCC_MAP),
                                 shadowState.ambientOcclusionMap,
                                 samplingProperties);
