+++ source/MaterialXTest/MaterialXRenderGlsl/RenderGlsl.cpp
@@ -6,6 +6,9 @@
 #include <MaterialXTest/External/Catch/catch.hpp>
 #include <MaterialXTest/MaterialXRender/RenderUtil.h>
 
+#include <MaterialXGenGlsl/GlslShaderGenerator.h>
+#include <MaterialXRenderGlsl/GlslRenderer.h>
+#include <MaterialXRenderGlsl/GLTextureHandler.h>
 #include <MaterialXRenderGlsl/TextureBaker.h>
 
 #include <MaterialXRender/GeometryHandler.h>
@@ -120,7 +123,7 @@ void GlslShaderRenderTester::createRenderer(std::ostream& log)
 
         // Set image handler on renderer
         mx::StbImageLoaderPtr stbLoader = mx::StbImageLoader::create();
-        mx::ImageHandlerPtr imageHandler = mx::GLTextureHandler::create(stbLoader);
+        mx::ImageHandlerPtr imageHandler = _renderer->createImageHandler(stbLoader);
 #if defined(MATERIALX_BUILD_OIIO)
         mx::OiioImageLoaderPtr oiioLoader = mx::OiioImageLoader::create();
         imageHandler->addLoader(oiioLoader);
@@ -149,226 +152,6 @@ void GlslShaderRenderTester::createRenderer(std::ostream& log)
 bool GlslShaderRenderTester::saveImage(const mx::FilePath& filePath, mx::ConstImagePtr image, bool verticalFlip) const
 {
     return _renderer->getImageHandler()->saveImage(filePath, image, verticalFlip);
-}
-
-// If these streams don't exist add them for testing purposes
-//
-void addAdditionalTestStreams(mx::MeshPtr mesh)
-{
-    size_t vertexCount = mesh->getVertexCount();
-    if (vertexCount < 1)
-    {
-        return;
-    }
-
-    const std::string TEXCOORD_STREAM0_NAME("i_" + mx::MeshStream::TEXCOORD_ATTRIBUTE + "_0");
-    mx::MeshStreamPtr texCoordStream1 = mesh->getStream(TEXCOORD_STREAM0_NAME);
-    mx::MeshFloatBuffer uv = texCoordStream1->getData();
-
-    const std::string TEXCOORD_STREAM1_NAME("i_" + mx::MeshStream::TEXCOORD_ATTRIBUTE + "_1");
-    mx::MeshFloatBuffer* texCoordData2 = nullptr;
-    if (!mesh->getStream(TEXCOORD_STREAM1_NAME))
-    {
-        mx::MeshStreamPtr texCoordStream2 = mx::MeshStream::create(TEXCOORD_STREAM1_NAME, mx::MeshStream::TEXCOORD_ATTRIBUTE, 1);
-        texCoordStream2->setStride(2);
-        texCoordData2 = &(texCoordStream2->getData());
-        texCoordData2->resize(vertexCount * 2);
-        mesh->addStream(texCoordStream2);
-    }
-
-    const std::string COLOR_STREAM0_NAME("i_" + mx::MeshStream::COLOR_ATTRIBUTE + "_0");
-    mx::MeshFloatBuffer* colorData1 = nullptr;
-    if (!mesh->getStream(COLOR_STREAM0_NAME))
-    {
-        mx::MeshStreamPtr colorStream1 = mx::MeshStream::create(COLOR_STREAM0_NAME, mx::MeshStream::COLOR_ATTRIBUTE, 0);
-        colorData1 = &(colorStream1->getData());
-        colorStream1->setStride(4);
-        colorData1->resize(vertexCount * 4);
-        mesh->addStream(colorStream1);
-    }
-
-    const std::string COLOR_STREAM1_NAME("i_" + mx::MeshStream::COLOR_ATTRIBUTE + "_1");
-    mx::MeshFloatBuffer* colorData2 = nullptr;
-    if (!mesh->getStream(COLOR_STREAM1_NAME))
-    {
-        mx::MeshStreamPtr colorStream2 = mx::MeshStream::create(COLOR_STREAM1_NAME, mx::MeshStream::COLOR_ATTRIBUTE, 1);
-        colorData2 = &(colorStream2->getData());
-        colorStream2->setStride(4);
-        colorData2->resize(vertexCount * 4);
-        mesh->addStream(colorStream2);
-    }
-
-    const std::string GEOM_INT_STREAM_NAME("i_" + mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE + "_geompropvalue_integer");
-    int32_t* geomIntData = nullptr;
-    if (!mesh->getStream(GEOM_INT_STREAM_NAME))
-    {
-        mx::MeshStreamPtr geomIntStream = mx::MeshStream::create(GEOM_INT_STREAM_NAME, mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE, 0);
-        geomIntStream->setStride(1);
-        geomIntStream->getData().resize(vertexCount);
-        mesh->addStream(geomIntStream);
-        // Float and int32 have same size.
-        geomIntData = reinterpret_cast<int32_t*>(geomIntStream->getData().data());
-    }
-
-    const std::string GEOM_FLOAT_STREAM_NAME("i_" + mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE + "_geompropvalue_float");
-    mx::MeshFloatBuffer* geomFloatData = nullptr;
-    if (!mesh->getStream(GEOM_FLOAT_STREAM_NAME))
-    {
-        mx::MeshStreamPtr geomFloatStream = mx::MeshStream::create(GEOM_FLOAT_STREAM_NAME, mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE, 1);
-        geomFloatData = &(geomFloatStream->getData());
-        geomFloatStream->setStride(1);
-        geomFloatData->resize(vertexCount);
-        mesh->addStream(geomFloatStream);
-    }
-
-    const std::string GEOM_VECTOR2_STREAM_NAME("i_" + mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE + "_geompropvalue_vector2");
-    mx::MeshFloatBuffer* geomVector2Data = nullptr;
-    if (!mesh->getStream(GEOM_VECTOR2_STREAM_NAME))
-    {
-        mx::MeshStreamPtr geomVector2Stream = mx::MeshStream::create(GEOM_VECTOR2_STREAM_NAME, mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE, 1);
-        geomVector2Data = &(geomVector2Stream->getData());
-        geomVector2Stream->setStride(2);
-        geomVector2Data->resize(vertexCount * 2);
-        mesh->addStream(geomVector2Stream);
-    }
-
-    const std::string GEOM_VECTOR3_STREAM_NAME("i_" + mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE + "_geompropvalue_vector3");
-    mx::MeshFloatBuffer* geomVector3Data = nullptr;
-    if (!mesh->getStream(GEOM_VECTOR3_STREAM_NAME))
-    {
-        mx::MeshStreamPtr geomVector3Stream = mx::MeshStream::create(GEOM_VECTOR3_STREAM_NAME, mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE, 1);
-        geomVector3Data = &(geomVector3Stream->getData());
-        geomVector3Stream->setStride(3);
-        geomVector3Data->resize(vertexCount * 3);
-        mesh->addStream(geomVector3Stream);
-    }
-
-    const std::string GEOM_VECTOR4_STREAM_NAME("i_" + mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE + "_geompropvalue_vector4");
-    mx::MeshFloatBuffer* geomVector4Data = nullptr;
-    if (!mesh->getStream(GEOM_VECTOR4_STREAM_NAME))
-    {
-        mx::MeshStreamPtr geomVector4Stream = mx::MeshStream::create(GEOM_VECTOR4_STREAM_NAME, mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE, 1);
-        geomVector4Data = &(geomVector4Stream->getData());
-        geomVector4Stream->setStride(4);
-        geomVector4Data->resize(vertexCount * 4);
-        mesh->addStream(geomVector4Stream);
-    }
-
-    const std::string GEOM_COLOR2_STREAM_NAME("i_" + mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE + "_geompropvalue_color2");
-    mx::MeshFloatBuffer* geomColor2Data = nullptr;
-    if (!mesh->getStream(GEOM_COLOR2_STREAM_NAME))
-    {
-        mx::MeshStreamPtr geomColor2Stream = mx::MeshStream::create(GEOM_COLOR2_STREAM_NAME, mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE, 1);
-        geomColor2Data = &(geomColor2Stream->getData());
-        geomColor2Stream->setStride(2);
-        geomColor2Data->resize(vertexCount * 2);
-        mesh->addStream(geomColor2Stream);
-    }
-
-    const std::string GEOM_COLOR3_STREAM_NAME("i_" + mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE + "_geompropvalue_color3");
-    mx::MeshFloatBuffer* geomColor3Data = nullptr;
-    if (!mesh->getStream(GEOM_COLOR3_STREAM_NAME))
-    {
-        mx::MeshStreamPtr geomColor3Stream = mx::MeshStream::create(GEOM_COLOR3_STREAM_NAME, mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE, 1);
-        geomColor3Data = &(geomColor3Stream->getData());
-        geomColor3Stream->setStride(3);
-        geomColor3Data->resize(vertexCount * 3);
-        mesh->addStream(geomColor3Stream);
-    }
-
-    const std::string GEOM_COLOR4_STREAM_NAME("i_" + mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE + "_geompropvalue_color4");
-    mx::MeshFloatBuffer* geomColor4Data = nullptr;
-    if (!mesh->getStream(GEOM_COLOR4_STREAM_NAME))
-    {
-        mx::MeshStreamPtr geomColor4Stream = mx::MeshStream::create(GEOM_COLOR4_STREAM_NAME, mx::MeshStream::GEOMETRY_PROPERTY_ATTRIBUTE, 1);
-        geomColor4Data = &(geomColor4Stream->getData());
-        geomColor4Stream->setStride(4);
-        geomColor4Data->resize(vertexCount * 4);
-        mesh->addStream(geomColor4Stream);
-    }
-
-    auto sineData = [](float uv, float freq){
-        const float PI = std::acos(-1.0f);
-        float angle = uv * 2 * PI * freq;
-        return std::sin(angle) / 2.0f + 1.0f;
-    };
-    if (!uv.empty())
-    {
-        for (size_t i = 0; i < vertexCount; i++)
-        {
-            const size_t i2 = 2 * i;
-            const size_t i21 = i2 + 1;
-            const size_t i3 = 3 * i;
-            const size_t i4 = 4 * i;
-
-            // Fake second set of texture coordinates
-            if (texCoordData2)
-            {
-                (*texCoordData2)[i2] = uv[i21];
-                (*texCoordData2)[i21] = uv[i2];
-            }
-            if (colorData1)
-            {
-                // Fake some colors
-                (*colorData1)[i4] = uv[i2];
-                (*colorData1)[i4 + 1] = uv[i21];
-                (*colorData1)[i4 + 2] = 1.0f;
-                (*colorData1)[i4 + 3] = 1.0f;
-            }
-            if (colorData2)
-            {
-                (*colorData2)[i4] = 1.0f;
-                (*colorData2)[i4 + 1] = uv[i2];
-                (*colorData2)[i4 + 2] = uv[i21];
-                (*colorData2)[i4 + 3] = 1.0f;
-            }
-            if (geomIntData)
-            {
-                geomIntData[i] = static_cast<int32_t>(uv[i21] * 5);
-            }
-            if (geomFloatData)
-            {
-                (*geomFloatData)[i] = sineData(uv[i21], 12.0f);
-            }
-            if (geomVector2Data)
-            {
-                (*geomVector2Data)[i2] = sineData(uv[i21], 6.0f);
-                (*geomVector2Data)[i21] = 0.0f;
-            }
-            if (geomVector3Data)
-            {
-                (*geomVector3Data)[i3] = 0.0f;
-                (*geomVector3Data)[i3 + 1] = sineData(uv[i21], 8.0f);
-                (*geomVector3Data)[i3 + 2] = 0.0f;
-            }
-            if (geomVector4Data)
-            {
-                (*geomVector4Data)[i4] = 0.0f;
-                (*geomVector4Data)[i4 + 1] = 0.0f;
-                (*geomVector4Data)[i4 + 2] = sineData(uv[i21], 10.0f);
-                (*geomVector4Data)[i4 + 3] = 1.0f;
-            }
-
-            if (geomColor2Data)
-            {
-                (*geomColor2Data)[i2] = sineData(uv[i2], 10.0f);
-                (*geomColor2Data)[i21] = 0.0f;
-            }
-            if (geomColor3Data)
-            {
-                (*geomColor3Data)[i3] = 0.0f;
-                (*geomColor3Data)[i3 + 1] = sineData(uv[i2], 8.0f);
-                (*geomColor3Data)[i3 + 2] = 0.0f;
-            }
-            if (geomColor4Data)
-            {
-                (*geomColor4Data)[i4] = 0.0f;
-                (*geomColor4Data)[i4 + 1] = 0.0f;
-                (*geomColor4Data)[i4 + 2] = sineData(uv[i2], 6.0f);
-                (*geomColor4Data)[i4 + 3] = 1.0f;
-            }
-        }
-    }
 }
 
 bool GlslShaderRenderTester::runRenderer(const std::string& shaderName,
@@ -641,7 +424,7 @@ void GlslShaderRenderTester::runBake(mx::DocumentPtr doc, const mx::FileSearchPa
     const unsigned bakeHeight = std::max(bakeOptions.resolution, maxImageSize.second);
 
     mx::Image::BaseType baseType = bakeOptions.hdr ? mx::Image::BaseType::FLOAT : mx::Image::BaseType::UINT8;
-    mx::TextureBakerPtr baker = mx::TextureBaker::create(bakeWidth, bakeHeight, baseType);
+    mx::TextureBakerPtr baker = mx::TextureBakerGlsl::create(bakeWidth, bakeHeight, baseType);
     baker->setupUnitSystem(doc);
     baker->setImageHandler(_renderer->getImageHandler());
     baker->setOptimizeConstants(true);
