+++ source/MaterialXView/ViewerMSL.mm
@@ -479,7 +479,7 @@ void Viewer::bakeTextures()
 
         // Construct a texture baker.
         mx::Image::BaseType baseType = _bakeHdr ? mx::Image::BaseType::FLOAT : mx::Image::BaseType::UINT8;
-        mx::TextureBakerMslPtr baker = mx::TextureBakerMsl::create(MTL(device), bakeWidth, bakeHeight, baseType);
+        mx::TextureBakerPtr baker = mx::TextureBakerMsl::create(bakeWidth, bakeHeight, baseType);
         baker->setupUnitSystem(_stdLib);
         baker->setDistanceUnit(_genContext.getOptions().targetDistanceUnit);
         baker->setAverageImages(_bakeAverage);
